server/requirements.txt
backcall==0.1.0
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
Deprecated==1.2.4
elasticsearch==6.3.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-OAuthlib==0.9.5
Flask-Session==0.3.1
flask-shell-ipython==0.4.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
idna==2.7
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyexcel==0.5.15
pyexcel-ezodf==0.3.4
pyexcel-io==0.5.20
pyexcel-ods3==0.5.3
pyexcel-odsr==0.5.2
pyexcel-text==0.2.7.1
pyexcel-xlsx==0.5.7
PyGithub==1.43.3
Pygments==2.3.1
PyJWT==1.7.0
python-pushover==0.4
regex==2018.11.22
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
smmap2==2.0.5
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
